+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `xstring.FromBytes([]byte)` and `xstring.ToBytes(string)` for fast conversion
+
 ## v3.51.2
 * Added `table/options.ReadFromSnapshot(bool)` option for `session.StreamReadTable()`